Display on-chain funds in ui
This commit implements very crude displaying of on chain c-lightning
funds.

It also adds a withdraw method to get funds out of c-lightning's wallet
and a method to get a new address to add funds to c-lightning.

Most of the css classes for the new components were just re-used, these
need to be changed later.