ui: add browserlist stuff via npm start
While trying to check if the dependency updates broke anything, I ran `npm start` in a container, and got an error if answering "no" to the following:

```
Would you like to add the defaults to your package.json? … no
As of react-scripts >=2 you must specify targeted browsers.
Please add a browserslist key to your package.json.
```

If we say "yes" instead, we get this diff, and hit another error:
```
✔ We're unable to detect target browsers.

Would you like to add the defaults to your package.json? … yes

When specified, "proxy" in package.json must be a string.
Instead, the type of "proxy" was "object".
Either remove "proxy" from package.json, or make it a string.
```