ui: Change proxy to be a string
With npm 9.6.6 (what we get by doing `apk add --no-cache npm` in a podman container), we get the following:

```
✔ We're unable to detect target browsers.

Would you like to add the defaults to your package.json? … yes

When specified, "proxy" in package.json must be a string.
Instead, the type of "proxy" was "object".
Either remove "proxy" from package.json, or make it a string.
```

So we edit the "proxy" key manually to make it a string here. ¯\_(ツ)_/¯